# AArch64 BMLA: Reduce temporary register usage