## What do these changes do?
This PR changes the order of the BMLA instructions reduce the usage of temporary registers. Since this changes the data dependency of instructions this PR requires a thorough review.

Register `v31` is now no longer used, do we need to change something [here](https://github.com/larq/compute-engine/blob/24d853fb964ff44204f60bfd6684ed0b8772dbf7/larq_compute_engine/core/bgemm_kernels_arm64.h#L444-L452) as well?

## How Has This Been Tested?
CI

## Benchmark Results
I ran benchmarks on my phone with `num_calls=250`:
| Model          | Threads | master (ms) | PR (ms)    |
| -------------- | ------- | ----------- | ---------- |
| QuickNet       | 1       | 24.0 ± 0.3  | 24.2 ± 0.3 |
| QuickNet       | 2       | 18.2 ± 0.3  | 18.0 ± 0.3 |
|                |         |             |            |
| QuickNet Large | 1       | 36.2 ± 0.4  | 35.3 ± 0.4 |
| QuickNet Large | 2       | 26.3 ± 0.3  | 26.3 ± 0.4 |
|                |         |             |            |
| QuickNet XL    | 1       | 64.4 ± 0.6  | 64.1 ± 0.3 |
| QuickNet XL    | 2       | 45.9 ± 0.4  | 45.5 ± 0.3 |

It looks like changing the order of the instructions doesn't introduce any slow down, but take the exact numbers with a grain of salt since my phone can have quite a high run variation.

## Related issue number
https://github.com/larq/compute-engine/pull/332#issuecomment-613387689