@lgeiger 2020-06-01
#350 Added a faster 8x4 kernel which will be used in most cases, so optimizing the register usage here won't really matter much and reduces the readability. I am closing this for now.